+++ packages/common/CHANGELOG.md
@@ -30,6 +30,12 @@
 
 - c2a13806: Added StateTags
 
+## 0.5.24
+
+### Patch Changes
+
+- 204b5479: updated document schemas
+
 ## 0.5.23
 
 ### Patch Changes